requirements.txt
alabaster==0.7.12
amqp==5.0.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astroid==2.9.3
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==21.12b0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==6.2
crispy-bootstrap5==0.6
cryptography==36.0.1
cssselect==1.1.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.11
django-admin-honeypot==1.1.0
django-allauth==0.47.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-celery-beat==2.2.1
django-ckeditor==6.2.0
django-countries-plus==1.3.1
django-coverage-plugin==2.0.2
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-htmlmin==0.11.0
django-imagekit==4.1.0
django-jet2==1.0.12
django-js-asset==1.2.2
django-mathfilters==1.0.0
django-model-utils==4.2.0
django-phonenumber-field==6.0.0
django-redis==5.2.0
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-timezone-field==4.2.3
docutils==0.17.1
executing==0.8.2
factory-boy==3.2.1
Faker==11.3.0
feedparser==5.2.1
filelock==3.4.2
flake8==4.0.1
flake8-isort==4.1.1
google-api-python-client==1.4.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.20.2
httptools==0.3.0
identify==2.4.4
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.0.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
kombu==5.2.3
lazy-object-proxy==1.7.1
livereload==2.6.3
loguru==0.5.3
lxml==4.7.1
markdown2==2.4.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
nodeenv==1.6.0
oauth2client==4.1.3
oauthlib==3.1.1
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
phonenumbers==8.12.39
pickleshare==0.7.5
pilkit==2.0
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.16.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.6
pyquery==1.4.3
pytest==6.2.5
pytest-django==4.5.2
pytest-sugar==0.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-magic==0.4.24
python-memcached==1.59
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
pyuca==1.2
PyYAML==6.0
readtime==1.1.1
redis==4.1.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.4.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.27
sqlparse==0.4.2
stack-data==0.1.4
termcolor==1.1.0
testfixtures==6.18.3
text-unidecode==1.3
toml==0.10.2
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
types-pytz==2021.3.4
types-PyYAML==6.0.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.0
uvloop==0.16.0
vine==5.0.0
virtualenv==20.13.0
watchdog==2.1.6
watchgod==0.7
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.7.0
